Replace inherited classes with pure objects to remove static generation of factory functions (#19)
This PR refactors each node's instantiation/initialization and replace inherited classes with pure objects to remove the static generation of factory functions, that is aiming for better readability and maintainability.

Previously nodes are initialized as a class instance via static-generated factory functions. This has ended up with poor visibility and a confusing development experience.

To resolve that issue, I've decided to initialize them as pure JavaScript objects without the inheritance of base class and factory functions. While this leads to (a bit of) less performance, I think this makes a much easier development experience.